Finished fixing bugs on direction vectors. Refactored checl_event function for rotation events.

+++ headers/utils.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   utils.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: bbellavi <bbellavi@student.42.fr>          +#+  +:+       +#+        */
+/*   By: tony <tony@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/25 03:18:15 by bbellavi          #+#    #+#             */
-/*   Updated: 2020/04/23 17:15:55 by bbellavi         ###   ########.fr       */
+/*   Updated: 2020/04/23 18:07:51 by tony             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,6 +16,6 @@
 #include <math.h>
 
 t_vec2	rotate(t_vec2 u, double theta, char clockwise);
-double	to_radians(int degrees);
+double	to_radians(double degrees);
 
 #endif
